Cargo/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-compat==1.0.15
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
psycopg2-binary==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
